Move JwsSigner to BearerDidSigner
