--- crates/crypto/src/jose/mod.rs
@@ -1,1 +1,0 @@
-pub mod jws_signer;
