+++ crates/crypto/src/lib.rs
@@ -1,3 +1,2 @@
-pub mod jose;
 pub mod key;
 pub mod key_manager;
